Style: 상수 값 코딩 컨벤션에 맞게 수정
외부 환경 변수의 경우 static final로 주입 불가
코딩 컨벤션에서 static 키워드가 안 붙은 변수는 멤버변수로 카멜케이스에 감지됨
외부 주입되는 환경 변수의 경우 static 변수로 하고, 일반 값은 static final로 처리

Ref #3
