--- src/main/java/org/example/o2o/common/dto/jwt/Payload.java
@@ -1,15 +1,0 @@
-package org.example.o2o.common.dto.jwt;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class Payload {
-	private String subject;
-
-	private Long id;
-	private String accountId;
-	private String role;
-	private String name;
-}
