+++ src/main/java/org/example/o2o/common/dto/jwt/TokenDto.java
@@ -42,4 +42,15 @@ public class TokenDto {
 				.build();
 		}
 	}
+
+	@Getter
+	@NoArgsConstructor
+	public class Payload {
+		private String subject;
+
+		private Long id;
+		private String accountId;
+		private String role;
+		private String name;
+	}
 }
